Cast size to int after rounding to gig

+++ testcases/cloud_user/ebs/block_dev_map_tests.py
@@ -223,7 +223,7 @@ class Block_Device_Mapping_Tests(EutesterTestCase):
         url = url or self.url
         volumes = []
         self.image_bytes = self.get_remote_file_size_via_http(url=url)
-        self.image_gigs = (math.ceil(float(self.image_bytes)/self.gig) or 1)
+        self.image_gigs = int(math.ceil(float(self.image_bytes)/self.gig) or 1))
 
         self.status('Attempting to launch instance store instance...')
         instance = self.tester.run_image(self.image,
